+++ CHANGES.rst
@@ -40,6 +40,9 @@ Major changes
   learner that works well on tabular data. :pr:`926` by :user:`Jérôme Dockès
   <jeromedockes>`.
 
+* The minimum supported python version is now 3.9
+  :pr:`939` by :user:`Jérôme Dockès <jeromedockes>`.
+
 Minor changes
 -------------